Fixed problem 4

+++ src/problem003.cpp
@@ -1,3 +1,4 @@
+// problem003.cpp
 #include <iostream>
 #include <vector>
 #include <tools.h>
@@ -5,7 +6,7 @@
 int main() {
     std::cout << "Solving Project Euler Problem 003" << std::endl;
     
-    long long v = 600851475143LL;
+    ull v = 600851475143ULL;
     std::vector<fterm> factors = factor(v);
     std::cout << factors.back().p << std::endl;
